NIFI-2271: - Fixing issue that was preventing an anonymous user from accessing the log in page.
