+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/api/AccessResource.java
@@ -20,7 +20,6 @@ import com.wordnik.swagger.annotations.Api;
 import com.wordnik.swagger.annotations.ApiOperation;
 import com.wordnik.swagger.annotations.ApiResponse;
 import com.wordnik.swagger.annotations.ApiResponses;
-import io.jsonwebtoken.JwtException;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.nifi.admin.service.AdministrationException;
 import org.apache.nifi.authentication.AuthenticationResponse;
@@ -29,42 +28,22 @@ import org.apache.nifi.authentication.LoginIdentityProvider;
 import org.apache.nifi.authentication.exception.IdentityAccessException;
 import org.apache.nifi.authentication.exception.InvalidLoginCredentialsException;
 import org.apache.nifi.authorization.AccessDeniedException;
-import org.apache.nifi.authorization.AuthorizationRequest;
-import org.apache.nifi.authorization.AuthorizationResult;
-import org.apache.nifi.authorization.AuthorizationResult.Result;
-import org.apache.nifi.authorization.Authorizer;
-import org.apache.nifi.authorization.RequestAction;
-import org.apache.nifi.authorization.UserContextKeys;
-import org.apache.nifi.authorization.resource.ResourceFactory;
 import org.apache.nifi.authorization.user.NiFiUser;
-import org.apache.nifi.authorization.user.NiFiUserDetails;
 import org.apache.nifi.authorization.user.NiFiUserUtils;
 import org.apache.nifi.util.FormatUtils;
+import org.apache.nifi.web.NiFiServiceFacade;
 import org.apache.nifi.web.api.dto.AccessConfigurationDTO;
-import org.apache.nifi.web.api.dto.AccessStatusDTO;
 import org.apache.nifi.web.api.entity.AccessConfigurationEntity;
-import org.apache.nifi.web.api.entity.AccessStatusEntity;
-import org.apache.nifi.web.security.InvalidAuthenticationException;
-import org.apache.nifi.web.security.ProxiedEntitiesUtils;
-import org.apache.nifi.web.security.UntrustedProxyException;
-import org.apache.nifi.web.security.jwt.JwtAuthenticationFilter;
-import org.apache.nifi.web.security.jwt.JwtAuthenticationProvider;
-import org.apache.nifi.web.security.jwt.JwtAuthenticationRequestToken;
+import org.apache.nifi.web.api.entity.CurrentUserEntity;
 import org.apache.nifi.web.security.jwt.JwtService;
 import org.apache.nifi.web.security.kerberos.KerberosService;
 import org.apache.nifi.web.security.otp.OtpService;
 import org.apache.nifi.web.security.token.LoginAuthenticationToken;
-import org.apache.nifi.web.security.token.NiFiAuthenticationToken;
 import org.apache.nifi.web.security.token.OtpAuthenticationToken;
-import org.apache.nifi.web.security.x509.X509AuthenticationProvider;
-import org.apache.nifi.web.security.x509.X509AuthenticationRequestToken;
-import org.apache.nifi.web.security.x509.X509CertificateExtractor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.preauth.x509.X509PrincipalExtractor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.Consumes;
@@ -77,9 +56,6 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.net.URI;
-import java.security.cert.X509Certificate;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -94,43 +70,12 @@ public class AccessResource extends ApplicationResource {
 
     private static final Logger logger = LoggerFactory.getLogger(AccessResource.class);
 
-    private X509CertificateExtractor certificateExtractor;
-    private X509AuthenticationProvider x509AuthenticationProvider;
-    private X509PrincipalExtractor principalExtractor;
-
     private LoginIdentityProvider loginIdentityProvider;
-    private JwtAuthenticationProvider jwtAuthenticationProvider;
     private JwtService jwtService;
     private OtpService otpService;
 
     private KerberosService kerberosService;
-
-    private Authorizer authorizer;
-
-    /**
-     * Authorizes access to the flow.
-     */
-    private boolean hasFlowAccess(final NiFiUser user) {
-        final Map<String,String> userContext;
-        if (!StringUtils.isBlank(user.getClientAddress())) {
-            userContext = new HashMap<>();
-            userContext.put(UserContextKeys.CLIENT_ADDRESS.name(), user.getClientAddress());
-        } else {
-            userContext = null;
-        }
-
-        final AuthorizationRequest request = new AuthorizationRequest.Builder()
-                .resource(ResourceFactory.getFlowResource())
-                .identity(user.getIdentity())
-                .anonymous(user.isAnonymous())
-                .accessAttempt(true)
-                .action(RequestAction.READ)
-                .userContext(userContext)
-                .build();
-
-        final AuthorizationResult result = authorizer.authorize(request);
-        return Result.Approved.equals(result.getResult());
-    }
+    private NiFiServiceFacade serviceFacade;
 
     /**
      * Retrieves the access configuration for this NiFi.
@@ -162,97 +107,24 @@ public class AccessResource extends ApplicationResource {
     }
 
     /**
-     * Gets the status the client's access.
+     * Retrieves the identity of the user making the request.
      *
-     * @param httpServletRequest the servlet request
-     * @return A accessStatusEntity
+     * @return An identityEntity
      */
     @GET
     @Consumes(MediaType.WILDCARD)
     @Produces(MediaType.APPLICATION_JSON)
-    @Path("")
+    @Path("current-user")
     @ApiOperation(
-            value = "Gets the status the client's access",
-            response = AccessStatusEntity.class
+            value = "Retrieves the user identity of the user making the request",
+            response = CurrentUserEntity.class
     )
-    @ApiResponses(
-            value = {
-                    @ApiResponse(code = 400, message = "NiFi was unable to complete the request because it was invalid. The request should not be retried without modification."),
-                    @ApiResponse(code = 401, message = "Unable to determine access status because the client could not be authenticated."),
-                    @ApiResponse(code = 403, message = "Unable to determine access status because the client is not authorized to make this request."),
-                    @ApiResponse(code = 409, message = "Unable to determine access status because NiFi is not in the appropriate state."),
-                    @ApiResponse(code = 500, message = "Unable to determine access status because an unexpected error occurred.")
-            }
-    )
-    public Response getAccessStatus(@Context HttpServletRequest httpServletRequest) {
+    public Response getCurrentUser() {
+        // create the response entity
+        final CurrentUserEntity entity = serviceFacade.getCurrentUser();
 
-        // only consider user specific access over https
-        if (!httpServletRequest.isSecure()) {
-            throw new IllegalStateException("User authentication/authorization is only supported when running over HTTPS.");
-        }
-
-        final AccessStatusDTO accessStatus = new AccessStatusDTO();
-
-        try {
-            final X509Certificate[] certificates = certificateExtractor.extractClientCertificate(httpServletRequest);
-
-            // if there is not certificate, consider a token
-            if (certificates == null) {
-                // look for an authorization token
-                final String authorization = httpServletRequest.getHeader(JwtAuthenticationFilter.AUTHORIZATION);
-
-                // if there is no authorization header, we don't know the user
-                if (authorization == null) {
-                    accessStatus.setStatus(AccessStatusDTO.Status.UNKNOWN.name());
-                    accessStatus.setMessage("No credentials supplied, unknown user.");
-                } else {
-                    try {
-                        // Extract the Base64 encoded token from the Authorization header
-                        final String token = StringUtils.substringAfterLast(authorization, " ");
-
-                        final JwtAuthenticationRequestToken jwtRequest = new JwtAuthenticationRequestToken(token, httpServletRequest.getRemoteAddr());
-                        final NiFiAuthenticationToken authenticationResponse = (NiFiAuthenticationToken) jwtAuthenticationProvider.authenticate(jwtRequest);
-                        final NiFiUser nifiUser = ((NiFiUserDetails) authenticationResponse.getDetails()).getNiFiUser();
-
-                        // set the user identity
-                        accessStatus.setIdentity(nifiUser.getIdentity());
-
-                        // attempt authorize to /flow
-                        accessStatus.setStatus(AccessStatusDTO.Status.ACTIVE.name());
-                        accessStatus.setMessage("You are already logged in.");
-                    } catch (JwtException e) {
-                        throw new InvalidAuthenticationException(e.getMessage(), e);
-                    }
-                }
-            } else {
-                try {
-                    final X509AuthenticationRequestToken x509Request = new X509AuthenticationRequestToken(
-                            httpServletRequest.getHeader(ProxiedEntitiesUtils.PROXY_ENTITIES_CHAIN), principalExtractor, certificates, httpServletRequest.getRemoteAddr());
-
-                    final NiFiAuthenticationToken authenticationResponse = (NiFiAuthenticationToken) x509AuthenticationProvider.authenticate(x509Request);
-                    final NiFiUser nifiUser = ((NiFiUserDetails) authenticationResponse.getDetails()).getNiFiUser();
-
-                    // set the user identity
-                    accessStatus.setIdentity(nifiUser.getIdentity());
-
-                    // attempt authorize to /flow
-                    accessStatus.setStatus(AccessStatusDTO.Status.ACTIVE.name());
-                    accessStatus.setMessage("You are already logged in.");
-                } catch (final IllegalArgumentException iae) {
-                    throw new InvalidAuthenticationException(iae.getMessage(), iae);
-                }
-            }
-        } catch (final UntrustedProxyException upe) {
-            throw new AccessDeniedException(upe.getMessage(), upe);
-        } catch (final AuthenticationServiceException ase) {
-            throw new AdministrationException(ase.getMessage(), ase);
-        }
-
-        // create the entity
-        final AccessStatusEntity entity = new AccessStatusEntity();
-        entity.setAccessStatus(accessStatus);
-
-        return generateOkResponse(entity).build();
+        // generate the response
+        return clusterContext(generateOkResponse(entity)).build();
     }
 
     /**
@@ -507,8 +379,9 @@ public class AccessResource extends ApplicationResource {
     }
 
     // setters
-    public void setAuthorizer(Authorizer authorizer) {
-        this.authorizer = authorizer;
+
+    public void setServiceFacade(NiFiServiceFacade serviceFacade) {
+        this.serviceFacade = serviceFacade;
     }
 
     public void setLoginIdentityProvider(LoginIdentityProvider loginIdentityProvider) {
@@ -519,24 +392,8 @@ public class AccessResource extends ApplicationResource {
         this.jwtService = jwtService;
     }
 
-    public void setJwtAuthenticationProvider(JwtAuthenticationProvider jwtAuthenticationProvider) {
-        this.jwtAuthenticationProvider = jwtAuthenticationProvider;
-    }
-
     public void setKerberosService(KerberosService kerberosService) {
         this.kerberosService = kerberosService;
-    }
-
-    public void setX509AuthenticationProvider(X509AuthenticationProvider x509AuthenticationProvider) {
-        this.x509AuthenticationProvider = x509AuthenticationProvider;
-    }
-
-    public void setPrincipalExtractor(X509PrincipalExtractor principalExtractor) {
-        this.principalExtractor = principalExtractor;
-    }
-
-    public void setCertificateExtractor(X509CertificateExtractor certificateExtractor) {
-        this.certificateExtractor = certificateExtractor;
     }
 
     public void setOtpService(OtpService otpService) {
