+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/resources/nifi-web-api-context.xml
@@ -353,11 +353,8 @@
         <property name="authorizer" ref="authorizer"/>
     </bean>
     <bean id="accessResource" class="org.apache.nifi.web.api.AccessResource" scope="singleton">
+        <property name="serviceFacade" ref="serviceFacade"/>
         <property name="loginIdentityProvider" ref="loginIdentityProvider"/>
-        <property name="x509AuthenticationProvider" ref="x509AuthenticationProvider"/>
-        <property name="certificateExtractor" ref="certificateExtractor"/>
-        <property name="principalExtractor" ref="principalExtractor"/>
-        <property name="jwtAuthenticationProvider" ref="jwtAuthenticationProvider"/>
         <property name="jwtService" ref="jwtService"/>
         <property name="otpService" ref="otpService"/>
         <property name="kerberosService" ref="kerberosService"/>
